requirements.txt
torch
numpy
einops
torchvision
tqdm
diffusers
accelerate
transformers
Pillow
wandb
git+https://github.com/openai/CLIP.git